==> 00_DataGenerator/01_B_DataModelling.ipynb <==
# Data Modelling

- Build Dataset for PandaPower from initial dataset

import os
import arrow
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

cwd = os.getcwd()
print(cwd)
main_fn = os.path.join(cwd, "data_modelling")
print(main_fn)

fn = os.path.join(main_fn, 'main_profiles.csv')
df_base = pd.read_csv(fn, sep=';', index_col='index')

df_base.head()

drop_cols = df_base.columns
df_base['load_h0_normed_MW'] = df_base['load_h0_normed_kW'] / 1000
df_base['load_g0_normed_MW'] = df_base['load_g0_normed_kW'] / 1000
df_base['load_l0_normed_MW'] = df_base['load_l0_normed_kW'] / 1000
df_base['gen_pv_normed_MW'] = df_base['gen_pv_normed_kW'] 
df_base['gen_wind_normed_MW'] = df_base['gen_wind_normed_kW']
df_base['gen_gas_normed_MW'] = df_base['gen_gas_normed_kW'] 

df_base.drop(drop_cols, axis=1, inplace=True)

df_base.head()

df_data = pd.DataFrame()

load_mapping = {
    #            0  1  2  3  4  5
    1: np.array([2, 1, 0, 0, 0, 1]),
    2: np.array([4, 2, 0, 1, 1, 0]),
    3: np.array([2, 0, 0, 0, 1, 0]),
    4: np.array([0, 1, 0, 0, 0, 0]),
    5: np.array([3, 0, 0, 2, 0, 0]),
    6: np.array([1, 1, 0, 0, 0, 0]),
    7: np.array([1, 2, 0, 0, 0, 0]),
    8: np.array([1, 0, 0, 1, 0, 0]),
    9: np.array([3, 0, 0, 0, 0, 0]),
    10: np.array([0, 0, 0, 0, 0, 0]),
    11: np.array([2, 0, 0, 0, 0, 0]),
    12: np.array([0, 0, 0, 0, 0, 0]),
    13: np.array([2, 0, 2, 0, 2, 0]),
    14: np.array([1, 0, 1, 0, 1, 0]),
}

fn = os.path.join(main_fn, 'load_mapping.npy')
np.save(fn, load_mapping) 

for bus, load in load_mapping.items():
    df_data['AGG_BUS_{}'.format(bus)] = np.dot(load, df_base.T)

df_data.set_index(df_base.index,inplace=True)

df_data.head()

## Compute Q

def compute_q(p, cos_phi=0.9, mode='inductive'):
    abs_q = p * np.tan(np.arccos(cos_phi))
    # inductive load 'consumes' reactive power
    if mode == 'inductive':
        return abs_q
    # capacitve load 'provides' reactive power
    elif mode == 'capacitve':
        return -1 * abs_q
    else:
        print('ERROR: Illegal mode: %s. Falling back to default (inductive).' % str(mode))
        return abs_q

cols = df_data.columns

dfs = {}
for col in cols:
    tmp = pd.DataFrame(index=df_data.index, columns=['name', 'p_mw', 'q_mvar'])
    tmp['p_mw'] = df_data[col]
    i = len(tmp['p_mw'])
    tmp['name'] = [col] * i
    dfs[col] = tmp

for key, val in dfs.items():
    tmp_q = []
    for val in val.iterrows():
        p = val[1]['p_mw']
        q = compute_q(p)
        tmp_q.append(q)
    dfs[key]['q_mvar'] = tmp_q

dfs['AGG_BUS_14'].head()

second_index = np.arange(0, len(cols))
index_array = [np.array(dfs['AGG_BUS_1'].index), second_index]

idx = pd.MultiIndex.from_product(index_array) 

df_main = pd.DataFrame(index=idx, columns=['name', 'p_mw', 'q_mvar'])

df_main.head()

p_mw = []
q_mw = []
names = []
for idx, data in df_main.groupby(level=0):
    for col in cols:
        p = dfs[col].loc[idx]['p_mw']
        q = dfs[col].loc[idx]['q_mvar']
        p_mw.append(p)
        q_mw.append(q)
        names.append(col)

print(len(p_mw))
print(len(q_mw))
print(len(names))
assert len(p_mw) == len(q_mw) == len(names)

df_main['p_mw'] = p_mw
df_main['q_mvar'] = q_mw
df_main['name'] = names

df_main.head(14)

df_main.index.levels[1]

h5_fn = os.path.join(main_fn, 'main_agg.h5')
print(h5_fn)
df_main.to_hdf(h5_fn, mode='w', key='df')
