==> 00_DataGenerator/02_GridModelling.ipynb <==
# Power Grid Modelling

import os
import arrow
import pandas as pd

import numpy as np
import matplotlib.pyplot as plt
import pandapower as pp
import pandapower as pp

from pandapower import networks
from pandapower import plotting as pp_plot

plt.style.use('tableau-colorblind10')

from GridExecutor import Executor

cwd = os.getcwd()
print(cwd)
data_fn = os.path.join(cwd, 'data_modelling')
grid_fn = os.path.join(cwd, 'grid_modelling')
fig_fn = os.path.join(cwd, 'figs')
print(data_fn)
print(grid_fn)

fn = os.path.join(data_fn, 'main_agg.h5')
all_data = pd.read_hdf(fn, key='df')

start_date = '2020-01-01 00:00:00'
end_date = '2022-12-31 23:45:00'

data = all_data.loc[start_date:end_date].copy()

timestamps = [x[0] for x in data.index]

timestamps = list(dict.fromkeys(timestamps))

data.tail()

## Actual Grid Mdf_dataing

print('PandaPower Version: {}'.format(pp.__version__))

grid = pp.networks.create_cigre_network_mv(with_der=False)

# drop the predefined load
grid.load = grid.load[0:0]

# run power flow
pp.runpp(grid)

## Define Load Mapping 

for bus_id in range(1,15):
    pp.create_load(grid, bus=bus_id, p_mw=0., q_mvar=0., name='AGG_BUS_%d' % bus_id, scaling=1., in_service=True, controllable=False)

grid.load

duration = (366 + 365 + 365) * 96
print('Duration: {} 1/4 hours'.format(duration))

assert len(timestamps) >= duration 

executor = Executor(grid)

### Simulate 2020, 2021 and 2022

for t in range(duration):
    # run simulation
    try:
        ts = timestamps[t]
        frame = data.loc[ts]
        grid = executor.step(frame)
                
        if t%10000 == 0:
            print('Curent step: {}'.format(t))

    except Exception as e:
        print('Exception!')
        print(e)

simulation_data = executor.get_simulation_data()

assert len(simulation_data) == len(timestamps)

simulation_data.head()

simulation_data.index = timestamps

simulation_data.head()

### Reduce simulation data

cols = [s for s in simulation_data.columns if "loading" in s]
print(cols)

simulation_data_reduced = simulation_data[cols].copy()

## Export Data

h5_fn_r = os.path.join(data_fn, 'simulation_data_reduced.h5')
print(h5_fn_r)
simulation_data_reduced.to_hdf(h5_fn_r, mode='w', key='df')

h5_fn = os.path.join(data_fn, 'simulation_data.h5')
print(h5_fn)
simulation_data.to_hdf(h5_fn, mode='w', key='df')

## Save CIGRE Net

# Save net as JSON
filename = os.path.join(grid_fn, 'CIGRE_net.json')
pp.to_json(grid,filename)
