==> 00_DataGenerator/03_A_AnomalyModelling.ipynb <==
# Anomaly Modelling

### Imports

import os
import arrow
import pandas as pd
import time 
import numpy as np
import matplotlib.pyplot as plt

plt.style.use('tableau-colorblind10')

from AnomalyGenerator import AnomalyGenerator

### Set configs

cwd = os.getcwd()
print(cwd)
data_fn = os.path.join(cwd, 'data_modelling')
fig_fn = os.path.join(cwd, 'figs')
print(data_fn)

## Import Data

h5_fn = os.path.join(data_fn, 'simulation_data_reduced.h5')
simulation_data = pd.read_hdf(h5_fn, key='df')

simulation_data.head()

## Add Column for labels

simulation_data['label'] = np.zeros(len(simulation_data.index))

## Select year 2021 to be anomalous

anomalous_data = simulation_data.loc['2022-01-01 00:00:00':'2022-12-31 23:45:00']

## Make Anomalies Great Again

### Anomaly Configs

len(anomalous_data)

outliers = {
    'collective': [{'feature': 'line_1_1_loading', 
                    'timestamps': [(1000,1222), (4560,4580), (30001, 30050), (15000, 15024)],
                    'factors': [400,1000, 1200, 300],
                   },
                   {'feature': 'line_5_6_loading',
                    'factors': [100, 200, 100, 450],
                    'timestamps': [(3,10), (24000, 24020), (11111, 11222), (30000, 32000)],
                   },
                   {'feature': 'trafo_0_loading',
                    'factors': [110, 120, 100, 450],
                    'timestamps': [(1000,1009), (2200, 2210), (11133, 11144), (32000, 32021)],
                   },
                  ],
    
    'point': [{'feature': 'line_2_3_loading',
               'timestamps': [10, 30, 10000, 14420, 17010, 28021, 34340],
               'factors': [400, 100, 120, 200, 300, 110, 120],
              },
              {'feature': 'line_11_14_loading',
               'timestamps': [100, 3000, 11200, 15520, 18020, 26041, 33333],
               'factors': [400, 1000, 300, 200, 101, 90000, 400],
              },
              {'feature': 'trafo_1_loading',
               'timestamps': [1200, 5000, 13200, 17520, 17820, 24041, 32333],
               'factors': [30, 100, 420, 220, 100, 800, 400],
              }
             ],
    
    'noise': [{'feature': 'line_7_8_loading',
               'timestamps': [(10, 30), (4000, 4020), (26666, 26670)],
               'factors': [10, 3, 5],
              }
             ]
}

adGenerator = AnomalyGenerator(anomalous_data,outliers)

pd.options.mode.chained_assignment = None
anomalous_data = adGenerator.make_anomalous()
pd.options.mode.chained_assignment = 'warn'

anomalous_data.sample(20)

## Export Data

h5_fn = os.path.join(data_fn, 'anomalous_data_y_2022_reduced.h5')
print(h5_fn)
anomalous_data.to_hdf(h5_fn, mode='w', key='df')
