==> 00_DataGenerator/04_A_ConceptDriftModelling.ipynb <==
# Concept Drift Modelling

import os
import arrow
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import pandapower as pp

plt.style.use('tableau-colorblind10')

from GridExecutor import Executor
from ConceptDrift import Drifter

## Set configs

cwd = os.getcwd()
print(cwd)
data_fn = os.path.join(cwd, 'data_modelling')
grid_fn = os.path.join(cwd, 'grid_modelling')
fig_fn = os.path.join(cwd, 'figs')
print(data_fn)
print(grid_fn)

pp.__version__

## Load Net from JSON

fn = os.path.join(grid_fn, 'CIGRE_net.json')
grid = pp.from_json(fn)

## Load Data

fn = os.path.join(data_fn, 'main_agg.h5')
all_data = pd.read_hdf(fn, key='df')
fn2 = os.path.join(data_fn, 'main.csv')
unscaled_data = pd.read_csv(fn2, sep=';', index_col='index')

start_date = '2023-01-01 00:00:00'
arrw_start = arrow.get(start_date)
end_date = '2023-12-31 23:45:00'
arrw_end = arrow.get(end_date)
res = arrw_end - arrw_start
duration_days = res.days + 1
print(duration_days)

data = all_data.loc[start_date:end_date].copy()
data_unscaled = unscaled_data.loc[start_date:end_date].copy()

### Load Load Mapping

fn = os.path.join(data_fn, 'load_mapping.npy')
init_load_mapping =  np.load(fn,allow_pickle='TRUE').item()

### Scale some columns since they are in kW and not MW

drop_cols = data_unscaled.columns
data_unscaled['load_h0_normed_MW'] = data_unscaled['load_h0_normed_kW'] / 1000
data_unscaled['load_g0_normed_MW'] = data_unscaled['load_g0_normed_kW'] / 1000
data_unscaled['load_l0_normed_MW'] = data_unscaled['load_l0_normed_kW'] / 1000
data_unscaled['gen_pv_normed_MW'] = data_unscaled['gen_pv_normed_kW'] 
data_unscaled['gen_wind_normed_MW'] = data_unscaled['gen_wind_normed_kW']
data_unscaled['gen_gas_normed_MW'] = data_unscaled['gen_gas_normed_kW'] 

data_unscaled.drop(drop_cols, axis=1, inplace=True)

timestamps = [x[0] for x in data.index]
timestamps = list(dict.fromkeys(timestamps))

grid.load

## Define Conept Drifs

manipulate_switch =  [
    {
        'switch_id': 4,
        'set_closed': True,
        'at_time_idx': 20,
        'until_time_idx': 80
    },
    {
        'switch_id': 4,
        'set_closed': False,
        'at_time_idx': 80
    },
    {
        'switch_id':2,
        'set_closed': True,
        'at_time_idx': 96 * 62,
        'until_time_idx': 96 * 96
    },
    {
        'switch_id':2,
        'set_closed': False,
        'at_time_idx': 96 * 96
    },
]

load_mapping = [
    {
        'bus_id': 1,
        'load_mapping': [3, 0, 2, 0, 1, 0],
        'at_time_idx': 96 * 24,
        'until_time_idx': 96 * 64
    },
    {
        'bus_id': 3, # three means -> 'AGG_BUS_3'
        'load_mapping': [2, 1, 0, 0, 1, 0],
        'at_time_idx': 96 * 100,
        'until_time_idx': -1
    },
    {
        'bus_id': 10, 
        'load_mapping': [2, 1, 1, 0, 1, 0],
        'at_time_idx': 31 * 97,
        'until_time_idx': 36 * 96 
    },
    {
        'bus_id': 11,
        'load_mapping': [2, 1, 0, 0, 2, 0],
        'at_time_idx': 245 * 97,
        'until_time_idx': -1 
    },
    
]

change_cos_phi = [
    {
        'bus_id': 1,
        'load_mapping': [3, 0, 2, 0, 1, 0], # if None, use initial load_mapping!
        'load_id_in_load_mapping': [0],
        'cos_phi': 0.85,
        'at_time_idx': 323,
        'until_time_idx': 432
    },
    {
        'bus_id': 4,
        'load_mapping': None, # if None, use initial load_mapping!
        'load_id_in_load_mapping': [0],
        'cos_phi': 0.89,
        'at_time_idx': 400,
        'until_time_idx': 400 + 96 * 14
    },
    {
        'bus_id': 13,
        'load_mapping': None, # if None, use initial load_mapping!
        'load_id_in_load_mapping': [0],
        'cos_phi': 0.94,
        'at_time_idx': 96 * 100,
        'until_time_idx': 96 * 131
    },
    {
        'bus_id': 8,
        'load_mapping': [2, 0, 4, 0, 1, 0], # if None, use initial load_mapping!
        'load_id_in_load_mapping': [0,2],
        'cos_phi': 0.94,
        'at_time_idx': 96 * 200,
        'until_time_idx': 96 * 250
    },
    {
        'bus_id': 14,
        'load_mapping': [1, 0, 2, 0, 1, 0], # if None, use initial load_mapping!
        'load_id_in_load_mapping': [0,4],
        'cos_phi': 0.96,
        'at_time_idx': 96 * 180,
        'until_time_idx': 96 * 220
    },
    {
        'bus_id': 6,
        'load_mapping': None, # if None, use initial load_mapping!
        'load_id_in_load_mapping': [0,1],
        'cos_phi': 0.84,
        'at_time_idx': 96 * 23,
        'until_time_idx': 96 * 42
    },
    
]

concept_drift = {'manipulate_switch': manipulate_switch, 'load_mapping': load_mapping, 'change_cos_phi': change_cos_phi}

duration = duration_days * 96
print('Duration: {} 1/4 hours'.format(duration))

drifter = Drifter(grid=grid, concept_drifts=concept_drift, load_mapping=init_load_mapping, unscaled_data=data_unscaled, timestamp_list=timestamps, agg_data=data.copy())
executor = Executor(grid)

labels = drifter.annotate_drifting_labels()

## Manipulate Load Mapping before simulation

drifted_data = drifter.manipulate_load_mapping()

drifted_data = drifter.manipulate_cos_phi()

## Execute Simulation

def do_concept_drift(t, grid):
    events_manipulate_switch = drifter.check_timestamps(t)

    if events_manipulate_switch[0]:
        for thing1 in events_manipulate_switch[1]:                
            grid = drifter.manipulate_switch(thing1)

    return grid

for t in range(duration):
    # run simulation
    try:
        grid = do_concept_drift(t, grid)
        
        executor.update_grid(grid)
        ts = timestamps[t]
        frame = data.loc[ts]
        grid = executor.step(frame)
        
        if t%10000 == 0:
            print('Curent step: {}'.format(t))

    except Exception as e:
        print('Exception in timestep: {}!'.format(t))
        print(e)

drifter.simulation_done = True

simulation_data = executor.get_simulation_data()

simulation_data['drift_labels'] = labels

assert len(timestamps) == len(simulation_data)

simulation_data.index = timestamps

simulation_data.iloc[10:25]

## Save data

h5_fn = os.path.join(data_fn, 'drifted_data_y_2023_more_cos_phi.h5')
print(h5_fn)
simulation_data.to_hdf(h5_fn, mode='w', key='df')

cols = [s for s in simulation_data.columns if "loading" in s]
cols.append("drift_labels")
print(cols)

simulation_data_reduced = simulation_data.copy()
simulation_data_reduced = simulation_data_reduced[cols]

h5_fn = os.path.join(data_fn, 'drifted_data_y_2023_reduced_more_cos_phi.h5')
print(h5_fn)
simulation_data_reduced.to_hdf(h5_fn, mode='w', key='df')
